requirements.txt
requests>=2.31.0
xmltodict>=0.13.0
python-magic>=0.4.27
rich>=13.7.0
